move SourceConfigurer to separate pkg
